Implementing user deletion and pytest user deletion unitary test

+++ src/database/repositories/project_repository.py
@@ -1,28 +1,35 @@
+from ast import Module
 from typing import List
 from beanie import PydanticObjectId
 from api_requests.project_requests import CreateProjectRequest, PatchProjectRequest
 from api_responses.project_responses import ProjectResponse
 from database.models.device_model import Device
-from database.models.module_model import Module
 from database.models.project_model import Project
+from database.models.sensor_reading_model import SensorReading
 from database.repositories.base_repository import BaseRepository
 from exceptions.project_exceptions import ProjectNotFoundException, UpdateProjectException
 
-class ProjectRepository(BaseRepository):
+class ProjectRepository(
+    BaseRepository[
+        Project,
+        ProjectResponse,
+        CreateProjectRequest,
+        PatchProjectRequest
+    ]):
     "Repository for project CRUD operations."
 
-    def __init__(self, project_model: Project):
-        self.project_model = project_model
+    def __init__(self):
+        super().__init__(Project)
 
-    async def get_by_id(self, user_id: PydanticObjectId, project_id: PydanticObjectId) -> ProjectResponse:
+    async def get(self, user_id: PydanticObjectId, project_id: PydanticObjectId) -> ProjectResponse:
         
-        project = await self.project_model.find_one(
-            self.project_model.user_id == user_id,
-            self.project_model.id == project_id
+        project = await self.model.find_one(
+            self.model.user_id == user_id,
+            self.model.id == project_id
         )
 
         if not project:
-            raise ProjectNotFoundException("Project not found.")
+            raise ProjectNotFoundException()
         
         modules = list(map(str, project.modules))
 
@@ -33,12 +40,29 @@ class ProjectRepository(BaseRepository):
             modules=modules
         )
     
+    async def get_all(self, user_id: PydanticObjectId) -> List[ProjectResponse]:
+
+        projects = await self.model.find(
+            self.model.user_id == user_id
+        ).to_list()
+
+        return [
+            ProjectResponse(
+                id=str(project.id),
+                name=project.name,
+                description=project.description,
+                modules=list(map(str, project.modules)),
+            )
+            for project in projects
+        ]
+    
     async def create(self, user_id: PydanticObjectId, create_project_request: CreateProjectRequest) -> ProjectResponse:
 
-        project = self.project_model(
+        project = self.model(
             user_id=user_id,
             name=create_project_request.name,
-            description=create_project_request.description
+            description=create_project_request.description,
+            modules=[]
         )
 
         await project.insert()
@@ -52,22 +76,6 @@ class ProjectRepository(BaseRepository):
             modules=modules
         )
     
-    async def list(self, user_id: PydanticObjectId) -> List[ProjectResponse]:
-
-        projects = await self.project_model.find(
-            self.project_model.user_id == user_id
-        ).to_list()
-
-        return [
-            ProjectResponse(
-                id=str(project.id),
-                name=project.name,
-                description=project.description,
-                modules=list(map(str, project.modules)),
-            )
-            for project in projects
-        ]
-    
     async def update(
             self, 
             user_id: PydanticObjectId, 
@@ -76,12 +84,12 @@ class ProjectRepository(BaseRepository):
         ) -> ProjectResponse:
         
         project = await self.project_model.find_one(
-            self.project_model.user_id == user_id,
-            self.project_model.id == project_id
+            self.model.user_id == user_id,
+            self.model.id == project_id
         )
 
         if not project:
-            raise ProjectNotFoundException("Project not found.")
+            raise ProjectNotFoundException()
 
         update_data = {key: value for key, value in patch_project_request.model_dump(exclude_unset=True).items()}
 
@@ -97,31 +105,41 @@ class ProjectRepository(BaseRepository):
             raise UpdateProjectException("Error while parsing update data.")
     
     async def delete(self, user_id: PydanticObjectId, project_id: PydanticObjectId):
-        
-        project = await self.project_model.find_one(
-            self.project_model.user_id == user_id,
-            self.project_model.id == project_id
+
+        project = await self.model.find_one(
+            self.model.user_id == user_id,
+            self.model.id == project_id
         )
 
         if not project:
             raise ProjectNotFoundException("Project not found.")
-        
-        modules = await Module.find(
-            Module.user_id == user_id,
-            Module.project_id == project_id
-        ).to_list()
 
-        for module in modules:
-            await Device.delete_many(
-                Device.module_id == module.id
-            )
+        if not project.modules:
+            return
 
-        await Module.delete_many(
-            Module.user_id == user_id,
-            Module.project_id == project_id
-        )
-        
+        module_ids = project.modules
+
+        device_ids = await Device.find(Device.module_id.in_(module_ids)).distinct(Device.id)
+
+        await SensorReading.delete_many(SensorReading.device_id.in_(device_ids))
+
+        await Device.delete_many(Device.module_id.in_(module_ids))
+
+        await Module.delete_many(Module.id.in_(module_ids))
+
         await project.delete()
+
+    async def exists(self, project_id: PydanticObjectId) -> bool:
+
+        project = await self.model.find_one(
+            self.model.id == project_id
+        )
+
+        if project:
+            return True
+        
+        return False
+
